agri-platform/backend/requirements.txt
flask
pymongo
flask-cors
python-dotenv
requests
gunicorn
